+++ examples/cpchop.cpp
@@ -31,7 +31,7 @@
 
 #include <opm/common/utility/platform_dependent/reenable_warnings.h>
 
-#include <opm/core/utility/MonotCubicInterpolator.hpp>
+#include <opm/common/utility/numeric/MonotCubicInterpolator.hpp>
 #include <opm/parser/eclipse/Units/Units.hpp>
 
 #include <opm/output/eclipse/EclipseGridInspector.hpp>
