+++ examples/upscale_cond.cpp
@@ -62,7 +62,7 @@
 
 #include <opm/common/utility/platform_dependent/reenable_warnings.h>
 
-#include <opm/core/utility/MonotCubicInterpolator.hpp>
+#include <opm/common/utility/numeric/MonotCubicInterpolator.hpp>
 #include <opm/parser/eclipse/Units/Units.hpp>
 
 #include <opm/upscaling/RelPermUtils.hpp>
