+++ opm/porsol/common/RockAnisotropicRelperm.hpp
@@ -38,7 +38,7 @@
 
 
 #include <dune/common/fvector.hh>
-#include <opm/core/utility/NonuniformTableLinear.hpp>
+#include <opm/common/utility/numeric/NonuniformTableLinear.hpp>
 #include <opm/porsol/common/Matrix.hpp>
 
 #include <fstream>
