+++ opm/porsol/euler/EulerUpstreamResidual.hpp
@@ -41,7 +41,7 @@
 #include <boost/unordered_map.hpp>
 
 #include <opm/common/utility/parameters/ParameterGroup.hpp>
-#include <opm/core/utility/SparseVector.hpp>
+#include <opm/common/utility/numeric/SparseVector.hpp>
 
 
 namespace Opm {
