+++ opm/porsol/euler/ImplicitCapillarity_impl.hpp
@@ -39,7 +39,7 @@
 #include <opm/common/ErrorMacros.hpp>
 #include <opm/core/utility/Average.hpp>
 #include <opm/parser/eclipse/Units/Units.hpp>
-#include <opm/core/utility/RootFinders.hpp>
+#include <opm/common/utility/numeric/RootFinders.hpp>
 #include <opm/core/utility/StopWatch.hpp>
 #include <dune/grid/common/Volumes.hpp>
 #include <opm/porsol/common/ReservoirPropertyFixedMobility.hpp>
